+++ backend/src/test/java/com/woowacourse/gongseek/article/domain/repository/ArticleTagRepositoryTest.java
@@ -25,7 +25,7 @@ import org.springframework.boot.test.autoconfigure.orm.jpa.TestEntityManager;
 
 @SuppressWarnings("NonAsciiCharacters")
 @RepositoryTest
-public class ArticleRepositoryCustomTest {
+public class ArticleTagRepositoryTest {
 
     private final Member member = new Member("slo", "hanull", "avatar.com");
 
@@ -33,7 +33,7 @@ public class ArticleRepositoryCustomTest {
     private ArticleRepository articleRepository;
 
     @Autowired
-    private ArticleRepositoryCustom articleRepositoryCustom;
+    private ArticleTagRepository articleTagRepository;
 
     @Autowired
     private MemberRepository memberRepository;
@@ -87,9 +87,9 @@ public class ArticleRepositoryCustomTest {
                 .collect(Collectors.toList());
         article.addTag(new Tags(tags));
 
-        boolean firstResult = articleRepositoryCustom.existsArticleByTagId(tagIds.get(0));
-        boolean secondResult = articleRepositoryCustom.existsArticleByTagId(tagIds.get(1));
-        boolean thirdResult = articleRepositoryCustom.existsArticleByTagId(999L);
+        boolean firstResult = articleTagRepository.existsArticleByTagId(tagIds.get(0));
+        boolean secondResult = articleTagRepository.existsArticleByTagId(tagIds.get(1));
+        boolean thirdResult = articleTagRepository.existsArticleByTagId(999L);
 
         assertAll(
                 () -> assertThat(firstResult).isTrue(),
@@ -116,7 +116,7 @@ public class ArticleRepositoryCustomTest {
         firstArticle.addTag(new Tags(tags));
         secondArticle.addTag(new Tags(tags));
 
-        Map<Long, List<String>> foundTags = articleRepositoryCustom.findTags(
+        Map<Long, List<String>> foundTags = articleTagRepository.findTags(
                 List.of(firstArticle.getId(), secondArticle.getId()));
         assertAll(
                 () -> assertThat(foundTags.get(firstArticle.getId()).containsAll(tags)),
